Fixed export to force dot as decimal separator
On my PC the export looked like this: float coeff[] = {0,0016453f,0,0000000f,-0,0026317f,-0,0023588f,0,0034406f,0,0079769f,0 ...
Problem was that some cultures use a comma as a decimal separator (e.g. 1.2 is written as 1,2) and it breaks the list of coefficients, which also uses commas to separate values.
I forced a specific (en-US) culture, so that floating point numbers always use a dot (".") as a decimal separator.